+++ R/binom_test.R
@@ -83,7 +83,7 @@ binom_test <- function(x, n, p = 0.5,
   if(length(x) == 2) n <- sum(x)
   results <- stats::binom.test(x, n, p, alternative, conf.level) %>%
     tidy() %>%
-    rename(p = .data$p.value) %>%
+    rename(p = "p.value") %>%
     add_significance("p") %>%
     add_columns(n = n, .before = 1)
   if(!detailed){
@@ -126,7 +126,7 @@ pairwise_binom_test <- function(x, p.adjust.method = "holm", alternative = "two.
     adjust_pvalue("p", method = p.adjust.method) %>%
     add_significance("p.adj") %>%
     mutate(p.adj = signif(.data$p.adj, digits = 3)) %>%
-    select(-.data$p.signif)
+    select(-any_of("p.signif"))
     # select(.data$group1, .data$group2, .data$p, .data$p.adj, .data$p.adj.signif)
   results %>%
     set_attrs(args = args) %>%
@@ -165,7 +165,7 @@ pairwise_binom_test_against_p <- function(x, p = rep(1/length(x), length(x)), p.
     adjust_pvalue("p", method = p.adjust.method) %>%
     add_significance("p.adj") %>%
     mutate(p.adj = signif(.data$p.adj, digits = 3)) %>%
-    select(-.data$p.signif) %>%
+    select(-any_of("p.signif")) %>%
     # select(.data$p, .data$p.adj, .data$p.adj.signif) %>%
     add_columns(group = groups, observed = x, expected = p*sum(x), .before = 1)
   results %>%